requirements.txt
openai==2.24.0
python-dotenv==1.2.1
pypdf==6.7.3
beautifulsoup4==4.14.3
requests==2.32.5
markdown==3.10.2
xhtml2pdf==0.2.17
